requirements.txt
textual>=0.47.0
PyGithub>=2.1.0
python-dotenv>=1.0.0
